requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
pyyaml==6.0.1
openpyxl==3.1.2

# Add your analysis dependencies here:
# scikit-learn==1.3.2
# matplotlib==3.8.2
# seaborn==0.13.0
# requests==2.31.0

